fix: Always feed PSK through LabeledExtract
Before, we only did this if len(psk) > Nb. This “pulled out” the call
to hash from inside HMAC, which hashes the key if the key is longer than
Nb. By hashing such keys outside of HMAC together with a prefix we
wanted to make sure that there is no collision of inputs.

All this effort is based on an indifferentiability theorem on HMAC [1],
which says that HMAC is indifferentiable from a (keyed) random oracle.
It turns out I did not read close enough under which assumptions this
theorem is proved. Theorem 4.4 [1] makes an assumption on the keys:

  Fix d,n>0 with d≥n. Let f: {0,1}^n × {0,1}^d → {0,1}^n be a RO and
  consider K-restricted HMAC_d[SMD[f]] for an allowed key set K.
  Let R: {0,1}^∗ × {0,1}^∗ → {0,1}^n be a keyed RO.

The allowed key set K is given on page 25 of [1]. Already, keys must be
of size <= Nb, and then adhere to some specific conditions, which are
too cumbersome to implement in a crypto library. The most
straightforward restriction of the allowed key set is:

   the set of all keys of a fixed length that is less than Nb−1.

By feeding all PSKs through LabeledExtract, regardless of their length,
we guarantee a fixed-length key. For SHA-256, SHA-384, and SHA-512,
BLAKE2s, BLAKE2b, we have Nh < Nb-1.

[1] https://eprint.iacr.org/2013/382.pdf
    Yevgeniy Dodis and Thomas Ristenpart and John Steinberger and
    Stefano Tessaro
    To Hash or Not to Hash Again? (In)differentiability Results for H^2
    and HMAC