feature: Only hash PSK in PSK and AuthPSK modes
In the Base and Auth modes, the only PSK ever used is the default psk,
which is constant and all zero. Instead of transforming the Nh bytes
default psk into another constant of length Nh bytes, using
LabeledExtract, we can just not transform the default psk at all and
use it directly. This removes the performance penalty for non-psk modes.